requirements.txt
yt-dlp 
spotipy
Flask
requests
python-dotenv
ffmpeg-python